requirements.txt
cached_property
networkx
penman
pytorch-ignite
regex
sacrebleu
smatch
transformers==4.11.0
wandb
PyYAML>=5.1
urllib3==1.26.6 
black==22.1.0
sentencepiece==0.1.96
